LPD-59710 Adding feature flag

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
@@ -696,8 +696,10 @@ public class ObjectEntryLocalServiceImpl
 			Collections.emptyMap(), objectDefinition.getObjectDefinitionId(),
 			objectEntry::getValues);
 
-		_trashEntryLocalService.deleteEntry(
-			objectDefinition.getClassName(), objectEntry.getPrimaryKey());
+		if (FeatureFlagManagerUtil.isEnabled("LPD-53981")) {
+			_trashEntryLocalService.deleteEntry(
+				objectDefinition.getClassName(), objectEntry.getPrimaryKey());
+		}
 
 		if (!ObjectDefinitionThreadLocal.isDeleteObjectDefinitionId(
 				objectDefinition.getObjectDefinitionId())) {
